Move org.scijava.jupyter.notebook to org.scijava.notebook
