+++ src/main/java/org/scijava/notebook/converter/MapToJSONNotebookConverter.java
@@ -13,20 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scijava.jupyter.notebook.converter;
+package org.scijava.notebook.converter;
 
+import java.util.Map;
 import org.scijava.Priority;
 import org.scijava.convert.Converter;
-import org.scijava.jupyter.notebook.converter.ouput.JSONNotebookOutput;
+import org.scijava.notebook.converter.ouput.JSONNotebookOutput;
 import org.scijava.plugin.Plugin;
 
 @Plugin(type = Converter.class, priority = Priority.LOW_PRIORITY)
-public class StringToJSONNotebookConverter<O extends String>
+public class MapToJSONNotebookConverter<O extends Map>
         extends NotebookOutputConverter<O, JSONNotebookOutput> {
 
     @Override
     public Class getInputType() {
-        return String.class;
+        return Map.class;
     }
 
     @Override
@@ -37,7 +38,7 @@ public class StringToJSONNotebookConverter<O extends String>
     @Override
     public JSONNotebookOutput convert(Object object) {
         return new JSONNotebookOutput(JSONNotebookOutput.getMimeType(),
-                (String) object);
+                (String) object.toString());
     }
 
 }
